Updated standard blendfunc
Kiwi.js has changed its standard blendfunc. This commit replaces old
blendfunc with the new version.

This is a stopgap measure. Future versions of Kiwi.js are planned to
implement a blend mode manager for WebGL renderers. This particle
plugin will need revision to remain compliant with the new renderer
functions, and will no longer need to restore a default.