Create 1.2.1 branch. Remove gl-matrix dependency.
This version is a necessary compatibility release with KiwiJS v1.3.1. We removed the gl-matrix library from KiwiJS, which was used for some camera matrix operations in this plugin. I've removed this dependency by using just KiwiJS matrix ops.

Also implement a better world angle update system. The prior version would only update when the particle shader was enabled. As this would happen if there was anything else in the state with the particle object, it did not occur often. However, I've fixed it for the rare but possible case that the user is experimenting with a particle effect and nothing else. This has been fixed: world angle uniform is updated per-frame, along with time and camera matrix uniforms.